requirements.txt
flask>=2.0.1
flask-cors>=3.0.10
python-dotenv>=0.19.0
requests>=2.26.0
pytz>=2021.3
python-dateutil>=2.8.2
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
google-api-python-client>=2.16.0
schedule>=1.1.0
pytest>=6.2.5
gunicorn>=21.2.0
pymongo>=4.0.0
PyJWT>=2.4.0
Authlib>=1.0.0
sendgrid>=6.9.0
Jinja2>=3.0.0
